# Task write endpoints on the task API
## Problem

The task write endpoints lived on the workflow-run API. Reviewing `setTaskState` on the way also surfaced two problems: it is a pessimistic write (the request carries no expected revision) but locked nothing, so concurrent calls for the same run could lose updates; and it hashed the task input inside the open transaction.

## Changes

**`workflowRun.transitionTaskStateV1` → `task.transitionStateV1`, `workflowRun.setTaskStateV1` → `task.setStateV1`.** The request and response types move to the task API module, and the set-state service logic moves from the workflow-run service to the task service.

```typescript
// before (workflowRun.transitionTaskStateV1)
{ type: "retry"; id: runId; taskId; expectedWorkflowRunRevision; taskState }

// after (task.transitionStateV1)
{ type: "retry"; id: taskId; workflowRunId; expectedWorkflowRunRevision; taskState }
```

**`setTaskState` locks the run row.** Since the request carries no expected revision, both branches begin their transaction by locking the run row, which serializes concurrent writers for the same run. Input hashing moved ahead of the transaction, and a new task's two transition rows are one batch insert.

**Run-row locks are narrowed.** The joined run-with-state read now locks only the run row (`FOR UPDATE OF workflow_run`), not the append-only transition row it joins, and the plain run read gains an optional `forUpdate` for callers that need only existence plus the lock.

## Breaking changes

- `transitionTaskStateV1` and `setTaskStateV1` moved from the workflowRun API to `task.transitionStateV1` / `task.setStateV1`; requests use `id` for the task and `workflowRunId` for the run.
- The transition/set-state request types moved from `@aikirun/types/workflow/task` to `@aikirun/types/api/task`.
